src/requirements.txt
boto3
PyGithub
requests
